requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.5
Django==5.2.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
jwt==1.4.0
Markdown==3.8.2
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
